refactor: moved classes and interfaces to their respective projects
