+++ Kata.Shared/Interfaces/IDestination.cs
@@ -1,0 +1,6 @@
+﻿namespace Kata.Shared.Interfaces;
+
+public interface IDestination
+{
+    void WriteChar(char charToCopy);
+}
