+++ Kata.Shared/Interfaces/ISource.cs
@@ -1,0 +1,5 @@
+﻿namespace Kata.Shared.Interfaces;
+public interface ISource
+{
+    char ReadChar();
+}
