+++ Kata.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kata.Tests", "Kata.Tests\Ka
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kata", "Kata\Kata.csproj", "{C2271D1E-DAC2-43AF-A53E-C0659A2C0E0C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kata.Shared", "Kata.Shared\Kata.Shared.csproj", "{376E8D80-78A7-4F81-BB8D-550AF1A0E287}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{C2271D1E-DAC2-43AF-A53E-C0659A2C0E0C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C2271D1E-DAC2-43AF-A53E-C0659A2C0E0C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C2271D1E-DAC2-43AF-A53E-C0659A2C0E0C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{376E8D80-78A7-4F81-BB8D-550AF1A0E287}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{376E8D80-78A7-4F81-BB8D-550AF1A0E287}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{376E8D80-78A7-4F81-BB8D-550AF1A0E287}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{376E8D80-78A7-4F81-BB8D-550AF1A0E287}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
