+++ Kata/Copier.cs
@@ -1,0 +1,11 @@
+﻿using Kata.Shared.Interfaces;
+
+namespace Kata;
+
+public class Copier(ISource source, IDestination destination)
+{
+    public void Copy()
+    {
+        
+    }
+}
